chore: created `Assessments.AssessmentsGridComponent`
removed a lot of duplicated code between the strand and moment assessment grids
