+++ lib/lanttern_web/live/pages/strands/id/assessment_component.ex
@@ -1,21 +1,10 @@
 defmodule LantternWeb.StrandLive.AssessmentComponent do
   use LantternWeb, :live_component
 
-  alias Lanttern.Assessments
-  alias Lanttern.Assessments.AssessmentPoint
-  alias Lanttern.Curricula.CurriculumItem
-  alias Lanttern.Filters
-  alias Lanttern.Identity.User
-  alias Lanttern.LearningContext.Moment
-  alias Lanttern.LearningContext.Strand
-  alias Lanttern.Schools.Student
-
-  import LantternWeb.AssessmentsHelpers, only: [save_entry_editor_component_changes: 2]
-  import LantternWeb.FiltersHelpers, only: [assign_user_filters: 4, assign_user_filters: 3]
+  import LantternWeb.FiltersHelpers, only: [assign_user_filters: 4]
 
   # shared components
-  alias LantternWeb.Assessments.EntryCellComponent
-  alias LantternWeb.Assessments.EntryDetailsComponent
+  alias LantternWeb.Assessments.AssessmentsGridComponent
   alias LantternWeb.StrandLive.StrandRubricsComponent
 
   @impl true
@@ -50,104 +39,17 @@ defmodule LantternWeb.StrandLive.AssessmentComponent do
             </p>
           <% end %>
         </div>
-        <div class="flex items-center gap-4 mt-6">
-          <.assessment_group_by_dropdow
-            current_assessment_group_by={@current_assessment_group_by}
-            myself={@myself}
-          />
-          <.assessment_view_dropdow
-            current_assessment_view={@current_assessment_view}
-            myself={@myself}
-          />
-        </div>
-        <%!-- if no assessment points, render empty state --%>
-        <div :if={!@has_assessment_points} class="p-10 mt-4 rounded shadow-xl bg-white">
-          <.empty_state><%= gettext("No assessment points for this strand yet") %></.empty_state>
-        </div>
-        <%!-- if no class filter is select, just render assessment points --%>
-        <div
-          :if={@selected_classes == [] && @has_assessment_points}
-          class="p-10 mt-4 rounded shadow-xl bg-white"
-        >
-          <p class="mb-6 font-bold text-ltrn-subtle"><%= gettext("Current assessment points") %></p>
-          <ol phx-update="stream" id="assessment-points-no-class" class="flex flex-col gap-4">
-            <.no_class_assessment_points_group
-              :for={{_dom_id, assessment_points_group} <- @streams.assessment_points}
-              assessment_points_group={assessment_points_group}
-            />
-          </ol>
-        </div>
       </.responsive_container>
-      <%!-- show entries only with class filter selected --%>
-      <div :if={@selected_classes != [] && @has_assessment_points} class="px-6">
-        <div class={[
-          "relative w-full max-h-[calc(100vh-4rem)] border mt-6 rounded shadow-xl #{@view_bg} overflow-x-auto",
-          if(@current_assessment_view == "student",
-            do: "border-ltrn-student-accent",
-            else: "border-transparent"
-          )
-        ]}>
-          <div
-            class="relative grid pb-4"
-            style={"grid-template-columns: 15rem #{@assessment_points_columns_grid}"}
-          >
-            <div
-              class={"sticky top-0 z-20 grid grid-cols-subgrid pt-4 #{@view_bg}"}
-              style={"grid-column: span #{@assessment_points_count + 1} / span #{@assessment_points_count + 1}"}
-            >
-              <div class={[
-                "sticky left-0 #{@view_bg}",
-                if(!is_nil(@current_assessment_group_by), do: "row-span-2")
-              ]}>
-              </div>
-              <div
-                :if={!is_nil(@current_assessment_group_by)}
-                id="grid-assessment-point-headers"
-                phx-update="stream"
-                class="grid grid-cols-subgrid"
-                style={"grid-column: span #{@assessment_points_count} / span #{@assessment_points_count}"}
-              >
-                <.assessment_point_header
-                  :for={{dom_id, ap_header} <- @streams.assessment_point_headers}
-                  id={dom_id}
-                  ap_header={ap_header}
-                  assessment_view={@current_assessment_view}
-                />
-              </div>
-              <div
-                id="grid-assessment-points"
-                phx-update="stream"
-                class="grid grid-cols-subgrid"
-                style={"grid-column: span #{@assessment_points_count} / span #{@assessment_points_count}"}
-              >
-                <.assessment_point
-                  :for={{dom_id, assessment_point} <- @streams.assessment_points}
-                  id={dom_id}
-                  assessment_point={assessment_point}
-                  assessment_view={@current_assessment_view}
-                />
-              </div>
-            </div>
-            <div
-              id="grid-student-entries"
-              phx-update="stream"
-              class="grid grid-cols-subgrid"
-              style={"grid-column: span #{@assessment_points_count + 1} / span #{@assessment_points_count + 1}"}
-            >
-              <.student_entries
-                :for={{dom_id, {student, entries}} <- @streams.students_entries}
-                id={dom_id}
-                student={student}
-                entries={entries}
-                myself={@myself}
-                current_assessment_view={@current_assessment_view}
-                view_bg={@view_bg}
-                current_user={@current_user}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
+      <.live_component
+        module={AssessmentsGridComponent}
+        id={:strand_assessment_grid}
+        current_user={@current_user}
+        strand_id={@strand.id}
+        classes_ids={@selected_classes_ids}
+        class="mt-6"
+        navigate={~p"/strands/#{@strand}?tab=assessment"}
+      />
+
       <.live_component
         module={StrandRubricsComponent}
         id={:strand_rubrics}
@@ -164,345 +66,6 @@ defmodule LantternWeb.StrandLive.AssessmentComponent do
         filter_opts={[strand_id: @strand.id]}
         navigate={~p"/strands/#{@strand}?tab=assessment"}
       />
-      <.fixed_bar :if={@entries_changes_map != %{}} class="flex items-center gap-6">
-        <div class="flex-1 flex items-center gap-4 text-sm">
-          <p class="text-white text-nowrap">
-            <%= ngettext("1 change", "%{count} changes", map_size(@entries_changes_map)) %>
-          </p>
-          <p
-            :if={@current_assessment_view == "student"}
-            class="flex items-center gap-2 font-bold text-ltrn-student-accent"
-          >
-            <.icon name="hero-information-circle" class="w-6 h-6" />
-            <%= gettext("You are registering students self-assessments") %>
-          </p>
-        </div>
-        <.button
-          phx-click={JS.navigate(~p"/strands/#{@strand}?tab=assessment")}
-          theme="ghost"
-          data-confirm={gettext("Are you sure?")}
-        >
-          <%= gettext("Discard") %>
-        </.button>
-        <.button
-          type="button"
-          phx-click="save_changes"
-          phx-target={@myself}
-          theme={if @current_assessment_view == "student", do: "student"}
-        >
-          <%= if @current_assessment_view == "student",
-            do: gettext("Save self-assessments"),
-            else: gettext("Save") %>
-        </.button>
-      </.fixed_bar>
-      <.slide_over
-        :if={@assessment_point_entry}
-        id="entry-details-overlay"
-        show={true}
-        on_cancel={JS.push("close_entry_details_overlay", target: @myself)}
-      >
-        <:title><%= gettext("Assessment point entry details") %></:title>
-        <.live_component
-          module={EntryDetailsComponent}
-          id={@assessment_point_entry.id}
-          entry={@assessment_point_entry}
-          current_user={@current_user}
-          notify_component={@myself}
-        />
-      </.slide_over>
-    </div>
-    """
-  end
-
-  # function components
-
-  attr :current_assessment_group_by, :string, required: true
-  attr :myself, Phoenix.LiveComponent.CID, required: true
-
-  def assessment_group_by_dropdow(assigns) do
-    text =
-      case assigns.current_assessment_group_by do
-        "curriculum" -> gettext("Show all, grouped by curriculum")
-        "moment" -> gettext("Show all, grouped by moment")
-        _ -> gettext("Show only goals assessments")
-      end
-
-    assigns = assign(assigns, :text, text)
-
-    ~H"""
-    <div class="relative">
-      <.badge_button id="group-by-dropdown-button" icon_name="hero-chevron-down-mini">
-        <%= @text %>
-      </.badge_button>
-      <.dropdown_menu id="group-by-dropdown" button_id="group-by-dropdown-button" z_index="30">
-        <:item
-          text={gettext("Show only goals assessments")}
-          on_click={JS.push("change_group_by", value: %{"group_by" => nil}, target: @myself)}
-        />
-        <:item
-          text={gettext("Show all, grouped by curriculum")}
-          on_click={JS.push("change_group_by", value: %{"group_by" => "curriculum"}, target: @myself)}
-        />
-        <:item
-          text={gettext("Show all, grouped by moment")}
-          on_click={JS.push("change_group_by", value: %{"group_by" => "moment"}, target: @myself)}
-        />
-      </.dropdown_menu>
-    </div>
-    """
-  end
-
-  attr :current_assessment_view, :string, required: true
-  attr :myself, Phoenix.LiveComponent.CID, required: true
-
-  def assessment_view_dropdow(assigns) do
-    {theme, text} =
-      case assigns.current_assessment_view do
-        "teacher" -> {"teacher", gettext("Assessed by teacher")}
-        "student" -> {"student", gettext("Assessed by students")}
-        "compare" -> {"primary", gettext("Compare teacher/students")}
-      end
-
-    assigns =
-      assigns
-      |> assign(:theme, theme)
-      |> assign(:text, text)
-
-    ~H"""
-    <div class="relative">
-      <.badge_button id="view-dropdown-button" icon_name="hero-chevron-down-mini" theme={@theme}>
-        <%= @text %>
-      </.badge_button>
-      <.dropdown_menu id="view-dropdown" button_id="view-dropdown-button" z_index="30">
-        <:item
-          text={gettext("Assessed by teacher")}
-          on_click={JS.push("change_view", value: %{"view" => "teacher"}, target: @myself)}
-        />
-        <:item
-          text={gettext("Assessed by students")}
-          on_click={JS.push("change_view", value: %{"view" => "student"}, target: @myself)}
-        />
-        <:item
-          text={gettext("Compare teacher and students assessments")}
-          on_click={JS.push("change_view", value: %{"view" => "compare"}, target: @myself)}
-        />
-      </.dropdown_menu>
-    </div>
-    """
-  end
-
-  attr :assessment_points_group, :any, required: true
-
-  def no_class_assessment_points_group(assigns) do
-    {_group_by_struct, assessment_points} = assigns.assessment_points_group
-    assigns = assign(assigns, :assessment_points, assessment_points)
-
-    ~H"""
-    <li :for={assessment_point <- @assessment_points} id={"no-class-#{assessment_point.id}"}>
-      <%= assessment_point.name %>
-    </li>
-    """
-  end
-
-  attr :id, :string, required: true
-  attr :ap_header, :any, required: true
-  attr :assessment_view, :string, required: true
-
-  def assessment_point_header(assigns) do
-    {header_struct, assessment_points_count} = assigns.ap_header
-
-    grid_column_span_style =
-      "grid-column: span #{assessment_points_count} / span #{assessment_points_count}"
-
-    assigns =
-      assigns
-      |> assign(:header_struct, header_struct)
-      |> assign(:grid_column_span_style, grid_column_span_style)
-
-    ~H"""
-    <div id={@id} class="group pt-2 px-2" style={@grid_column_span_style}>
-      <div class="h-full pb-2 border-b border-ltrn-light" style={@grid_column_span_style}>
-        <.assessment_point_header_struct header_struct={@header_struct} />
-      </div>
-    </div>
-    """
-  end
-
-  attr :header_struct, :any, required: true, doc: "moment, strand, or curriculum item"
-
-  def assessment_point_header_struct(%{header_struct: %Moment{}} = assigns) do
-    ~H"""
-    <.link
-      class="flex items-center w-full h-full p-1 rounded text-sm font-display font-bold truncate hover:bg-ltrn-mesh-cyan"
-      navigate={~p"/strands/moment/#{@header_struct.id}?tab=assessment"}
-    >
-      <%= @header_struct.name %>
-    </.link>
-    """
-  end
-
-  def assessment_point_header_struct(%{header_struct: %CurriculumItem{}} = assigns) do
-    ~H"""
-    <div class="flex items-center gap-2">
-      <.badge class="truncate">
-        <%= @header_struct.curriculum_component.name %>
-      </.badge>
-      <.badge :if={@header_struct.is_differentiation} theme="diff">
-        <%= gettext("Diff") %>
-      </.badge>
-    </div>
-    <p class="mt-1 text-sm line-clamp-2" title={@header_struct.name}>
-      <%= @header_struct.name %>
-    </p>
-    """
-  end
-
-  def assessment_point_header_struct(%{header_struct: %Strand{}} = assigns) do
-    ~H"""
-    <p class="flex items-center h-full text-sm font-display font-bold">
-      <%= gettext("Goals assessment") %>
-    </p>
-    """
-  end
-
-  attr :assessment_point, AssessmentPoint, required: true
-  attr :assessment_view, :string, required: true
-  attr :id, :string, required: true
-
-  def assessment_point(assigns) do
-    ~H"""
-    <div id={@id} class="flex flex-col p-2">
-      <div class="flex-1">
-        <.assessment_point_struct assessment_point={@assessment_point} />
-      </div>
-      <.compare_header :if={@assessment_view == "compare"} />
-    </div>
-    """
-  end
-
-  attr :assessment_point, AssessmentPoint, required: true
-
-  def assessment_point_struct(
-        %{assessment_point: %{curriculum_item: %CurriculumItem{}, moment_id: moment_id}} = assigns
-      )
-      when not is_nil(moment_id) do
-    tooltip =
-      """
-      #{assigns.assessment_point.name}
-
-      (#{assigns.assessment_point.curriculum_item.curriculum_component.name}) #{assigns.assessment_point.curriculum_item.name}
-      """
-
-    assigns = assign(assigns, :tooltip, tooltip)
-
-    ~H"""
-    <div class="flex flex-col">
-      <div :if={@assessment_point.is_differentiation} class="mb-1">
-        <.badge theme="diff"><%= gettext("Diff") %></.badge>
-      </div>
-      <p
-        class={[
-          "flex-1 text-sm",
-          if(@assessment_point.is_differentiation, do: "line-clamp-2", else: "line-clamp-3")
-        ]}
-        title={@tooltip}
-      >
-        <%= @assessment_point.name %>
-      </p>
-    </div>
-    """
-  end
-
-  def assessment_point_struct(
-        %{assessment_point: %{curriculum_item: %CurriculumItem{}}} = assigns
-      ) do
-    ~H"""
-    <div class="flex flex-col">
-      <div class="flex items-center gap-2">
-        <.badge class="truncate">
-          <%= @assessment_point.curriculum_item.curriculum_component.name %>
-        </.badge>
-        <.badge :if={@assessment_point.is_differentiation} theme="diff">
-          <%= gettext("Diff") %>
-        </.badge>
-      </div>
-      <p class="flex-1 mt-1 text-sm line-clamp-2" title={@assessment_point.curriculum_item.name}>
-        <%= @assessment_point.curriculum_item.name %>
-      </p>
-    </div>
-    """
-  end
-
-  def assessment_point_struct(%{assessment_point: %{moment: %Moment{}}} = assigns) do
-    ~H"""
-    <div class="text-sm whitespace-nowrap">
-      <.link
-        class="block w-full p-1 rounded overflow-hidden hover:bg-ltrn-mesh-cyan"
-        title={"#{@assessment_point.moment.name}\n\n#{@assessment_point.name}"}
-        navigate={~p"/strands/moment/#{@assessment_point.moment.id}?tab=assessment"}
-      >
-        <span class="font-bold"><%= @assessment_point.moment.name %></span> <br />
-        <span class="text-xs"><%= @assessment_point.name %></span>
-      </.link>
-    </div>
-    """
-  end
-
-  def assessment_point_struct(%{assessment_point: %{strand_id: strand_id}} = assigns)
-      when not is_nil(strand_id) do
-    ~H"""
-    <div class="p-1 text-sm whitespace-nowrap overflow-hidden">
-      <span class="font-bold"><%= gettext("Goal assessment") %></span> <br />
-      <span class="text-xs"><%= gettext("(Strand final assessment)") %></span>
-    </div>
-    """
-  end
-
-  def compare_header(assigns) do
-    ~H"""
-    <div class="flex gap-1 w-full mt-2">
-      <div class="flex-1 pb-1 border-b-2 border-ltrn-teacher-accent text-xs text-center text-ltrn-teacher-dark">
-        <%= gettext("Teacher") %>
-      </div>
-      <div class="flex-1 pb-1 border-b-2 border-ltrn-student-accent text-xs text-center text-ltrn-student-dark">
-        <%= gettext("Student") %>
-      </div>
-    </div>
-    """
-  end
-
-  attr :id, :string, required: true
-  attr :student, Student, required: true
-  attr :entries, :list, required: true
-  attr :myself, Phoenix.LiveComponent.CID, required: true
-  attr :current_assessment_view, :string, required: true
-  attr :view_bg, :string, required: true
-  attr :current_user, User, required: true
-
-  def student_entries(assigns) do
-    ~H"""
-    <div
-      id={@id}
-      class="grid grid-cols-subgrid"
-      style={"grid-column: span #{length(@entries) + 1} / span #{length(@entries) + 1}"}
-    >
-      <div class={"sticky left-0 z-10 pl-6 py-2 pr-2 #{@view_bg}"}>
-        <.profile_icon_with_name
-          profile_name={@student.name}
-          extra_info={@student.classes |> Enum.map(& &1.name) |> Enum.join(", ")}
-        />
-      </div>
-      <div :for={entry <- @entries} class="p-2">
-        <.live_component
-          module={EntryCellComponent}
-          id={"student-#{@student.id}-entry-for-#{entry.assessment_point_id}"}
-          class="w-full h-full"
-          entry={entry}
-          view={@current_assessment_view}
-          allow_edit={entry.is_strand_entry}
-          notify_component={@myself}
-        />
-      </div>
     </div>
     """
   end
@@ -510,260 +73,14 @@ defmodule LantternWeb.StrandLive.AssessmentComponent do
   # lifecycle
 
   @impl true
-  def mount(socket) do
-    socket =
-      socket
-      |> assign(:entries_changes_map, %{})
-      |> assign(:assessment_point_entry, nil)
-      |> assign(:has_entry_details_change, false)
-      |> stream_configure(
-        :assessment_point_headers,
-        dom_id: fn
-          {%CurriculumItem{} = ci, _count} -> "ap-group-curriculum-item-#{ci.id}"
-          {%Moment{} = moment, _count} -> "ap-group-moment-#{moment.id}"
-          {%Strand{} = strand, _count} -> "ap-group-strand-#{strand.id}"
-          _ -> ""
-        end
-      )
-      |> stream_configure(
-        :students_entries,
-        dom_id: fn {student, _entries} -> "student-#{student.id}" end
-      )
-
-    {:ok, socket}
-  end
-
-  @impl true
-  def update(
-        %{action: {EntryCellComponent, {:change, :cancel, composite_id, _, _}}},
-        socket
-      ) do
-    socket =
-      socket
-      |> update(:entries_changes_map, fn entries_changes_map ->
-        entries_changes_map
-        |> Map.drop([composite_id])
-      end)
-
-    {:ok, socket}
-  end
-
-  def update(
-        %{action: {EntryCellComponent, {:change, type, composite_id, entry_id, params}}},
-        socket
-      ) do
-    socket =
-      socket
-      |> update(:entries_changes_map, fn entries_changes_map ->
-        entries_changes_map
-        |> Map.put(composite_id, {type, entry_id, params})
-      end)
-
-    {:ok, socket}
-  end
-
-  def update(
-        %{action: {EntryCellComponent, {:view_details, entry}}},
-        socket
-      ) do
-    socket =
-      socket
-      |> assign(:assessment_point_entry, entry)
-
-    {:ok, socket}
-  end
-
-  def update(
-        %{action: {EntryDetailsComponent, {msg_type, _}}},
-        socket
-      )
-      when msg_type in [:change, :created_attachment, :deleted_attachment] do
-    {:ok, assign(socket, :has_entry_details_change, true)}
-  end
-
-  def update(
-        %{action: {EntryDetailsComponent, {:delete, _entry}}},
-        socket
-      ) do
-    socket =
-      socket
-      |> assign(:assessment_point_entry, nil)
-      |> stream_assessment_points()
-      |> stream_students_entries()
-      |> assign(:has_entry_details_change, false)
-
-    {:ok, socket}
-  end
-
   def update(%{strand: strand} = assigns, socket) do
     socket =
       socket
       |> assign(assigns)
       |> assign_user_filters([:classes], assigns.current_user, strand_id: strand.id)
-      |> assign_user_filters([:assessment_view], assigns.current_user)
-      |> assign_user_filters([:assessment_group_by], assigns.current_user)
-      |> assign_view_bg()
-      |> stream_assessment_points()
-      |> stream_students_entries()
 
     {:ok, socket}
   end
 
   def update(_assigns, socket), do: {:ok, socket}
-
-  defp assign_view_bg(socket) do
-    view_bg =
-      case socket.assigns.current_assessment_view do
-        "student" -> "bg-ltrn-student-lightest"
-        _ -> "bg-white"
-      end
-
-    assign(socket, :view_bg, view_bg)
-  end
-
-  defp stream_assessment_points(socket) do
-    {assessment_point_headers, assessment_points} =
-      Assessments.list_strand_assessment_points(
-        socket.assigns.strand.id,
-        socket.assigns.current_assessment_group_by
-      )
-
-    assessment_points_count = length(assessment_points)
-
-    assessment_points_columns_grid =
-      case socket.assigns.current_assessment_view do
-        "compare" ->
-          "repeat(#{assessment_points_count}, 12rem)"
-
-        _ ->
-          assessment_points
-          |> Enum.map(fn
-            %{strand_id: id} when not is_nil(id) -> "15rem"
-            _ -> "6rem"
-          end)
-          |> Enum.join(" ")
-      end
-
-    socket
-    |> stream(:assessment_points, assessment_points)
-    |> stream(:assessment_point_headers, assessment_point_headers)
-    |> assign(:assessment_points_count, length(assessment_points))
-    |> assign(:assessment_points_columns_grid, assessment_points_columns_grid)
-    |> assign(:has_assessment_points, assessment_points != [])
-  end
-
-  defp stream_students_entries(socket) do
-    students_entries =
-      Assessments.list_strand_students_entries(
-        socket.assigns.strand.id,
-        socket.assigns.current_assessment_group_by,
-        classes_ids: socket.assigns.selected_classes_ids,
-        check_if_has_evidences: true
-      )
-
-    socket
-    |> stream(:students_entries, students_entries)
-  end
-
-  # event handlers
-
-  @impl true
-  def handle_event(
-        "change_group_by",
-        %{"group_by" => group_by},
-        %{assigns: %{current_assessment_group_by: current_assessment_group_by}} = socket
-      )
-      when group_by == current_assessment_group_by,
-      do: {:noreply, socket}
-
-  def handle_event("change_group_by", %{"group_by" => group_by}, socket) do
-    # TODO
-    # before applying the group_by change, check if there're pending changes
-
-    Filters.set_profile_current_filters(
-      socket.assigns.current_user,
-      %{assessment_group_by: group_by}
-    )
-    |> case do
-      {:ok, _} ->
-        socket =
-          socket
-          |> assign(:current_assessment_group_by, group_by)
-          |> push_navigate(to: ~p"/strands/#{socket.assigns.strand}?tab=assessment")
-
-        {:noreply, socket}
-
-      {:error, _} ->
-        # do something with error?
-        {:noreply, socket}
-    end
-  end
-
-  def handle_event(
-        "change_view",
-        %{"view" => view},
-        %{assigns: %{current_assessment_view: current_assessment_view}} = socket
-      )
-      when view == current_assessment_view,
-      do: {:noreply, socket}
-
-  def handle_event("change_view", %{"view" => view}, socket) do
-    # TODO
-    # before applying the view change, check if there're pending changes
-
-    Filters.set_profile_current_filters(
-      socket.assigns.current_user,
-      %{assessment_view: view}
-    )
-    |> case do
-      {:ok, _} ->
-        socket =
-          socket
-          |> assign(:current_assessment_view, view)
-          |> push_navigate(to: ~p"/strands/#{socket.assigns.strand}?tab=assessment")
-
-        {:noreply, socket}
-
-      {:error, _} ->
-        # do something with error?
-        {:noreply, socket}
-    end
-  end
-
-  def handle_event("save_changes", _params, socket) do
-    %{
-      entries_changes_map: entries_changes_map,
-      current_user: current_user
-    } = socket.assigns
-
-    socket =
-      case save_entry_editor_component_changes(
-             entries_changes_map,
-             current_user.current_profile_id
-           ) do
-        {:ok, msg} -> put_flash(socket, :info, msg)
-        {:error, msg} -> put_flash(socket, :error, msg)
-      end
-      |> push_navigate(to: ~p"/strands/#{socket.assigns.strand}?tab=assessment")
-
-    {:noreply, socket}
-  end
-
-  def handle_event("close_entry_details_overlay", _, socket) do
-    socket =
-      socket
-      |> assign(:assessment_point_entry, nil)
-
-    socket =
-      if socket.assigns.has_entry_details_change do
-        socket
-        |> stream_assessment_points()
-        |> stream_students_entries()
-        |> assign(:has_entry_details_change, false)
-      else
-        socket
-      end
-
-    {:noreply, socket}
-  end
 end
